+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* Fixed: Remove nested links when showing content in flex loop.
 * Fixed: Remove redundent conditional checks in flex loop.
 * Fixed: Better archive setting check for Woo custom taxos.
 * Fixed: Extra entry pagination margin.